Update and rename Project-1.py to Word Puzzle Game
This Project is about Word Puzzle Game in which the user has to form the correct word out of a given set of characters. 
In the game the user has to solve this puzzle for 5 words, one at a time.
Problem words are stored in a list and appear to the user in random sequence.
Give the user score +1 for each correct answer and give -1 for each wrong answer. At last print the final score. 
You can store any number of words in the list, but in each round of the game only 5 words are shown to the user.